+++ src/lib/words/case.js
@@ -24,8 +24,11 @@ export function fixCase(string, locale) {
 
 
 
-	/* [2] Swapped cases (n-letter cases, i.e. uPPERCASE) */
-	pattern = "(\\b)(["+ locale.lowercaseChars +"])(["+ locale.uppercaseChars +"]{2,})";
+	/* [2] Swapped cases (n-letter cases, i.e. uPPERCASE)
+			There is one exception excluded from swapping → iOS
+			If needed, update this part of code to exclude the list of other popular names.
+	*/
+	pattern = "(\\b)(?!iOS)(["+ locale.lowercaseChars +"])(["+ locale.uppercaseChars +"]{2,})";
 	re = new RegExp(pattern, "g");
 	string = string.replace(re, function(string, $1, $2, $3){
 		return ($1 + $2 + $3.toLowerCase());